@@ -1,14 +1,21 @@
 {
   "name": "@cenk1cenk2/eslint-config",
-  "version": "2.7.59",
+  "version": "3.1.6",
   "description": "Eslint configuration for my taste.",
   "main": "index.js",
   "repository": "git@gitlab.kilic.dev:config/eslint-config.git",
   "author": "Cenk Kilic",
   "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": "./index.js",
+    "./vue": "./vue.js",
+    "./prettierrc": "./.prettierrc.js"
+  },
   "scripts": {
-    "format": "prettier --loglevel warn --write ./*.js && eslint --ext .ts,.js,.tsx,.jsx --fix ./",
-    "lint": "eslint --ext .ts,.js,.tsx,.jsx ./"
+    "format": "prettier --log-level warn --write ./*.js && eslint --fix",
+    "lint": "eslint",
+    "inspect": "eslint --inspect-config"
   },
   "simple-git-hooks": {
     "pre-commit": "pnpm exec lint-staged",
@@ -24,19 +31,19 @@
     ]
   },
   "peerDependencies": {
-    "@nuxtjs/eslint-config": ">= 7",
-    "@nuxtjs/eslint-config-typescript": ">= 7",
-    "eslint": ">= 8",
+    "@vue/eslint-config-typescript": ">= 14.1.1",
+    "eslint": ">= 9",
     "eslint-config-react": ">= 1",
     "eslint-plugin-react-hooks": ">= 4",
     "eslint-plugin-svelte": ">= 2",
+    "eslint-plugin-vue": ">= 9",
     "svelte-eslint-parser": ">= 0.26.0"
   },
   "peerDependenciesMeta": {
-    "@nuxtjs/eslint-config": {
+    "@vue/eslint-config-typescript": {
       "optional": true
     },
-    "@nuxtjs/eslint-config-typescript": {
+    "eslint-plugin-vue": {
       "optional": true
     },
     "eslint-config-react": {
@@ -53,17 +60,23 @@
     }
   },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
+    "@eslint/js": "^9.14.0",
+    "@stylistic/eslint-plugin": "^2.10.1",
+    "@typescript-eslint/eslint-plugin": "^8.14.0",
+    "@typescript-eslint/parser": "^8.14.0",
     "eslint-plugin-import": "^2.31.0",
-    "load-tsconfig": "^0.2.5"
+    "load-tsconfig": "^0.2.5",
+    "typescript-eslint": "^8.14.0"
   },
   "devDependencies": {
     "@cenk1cenk2/cz-cc": "^1.7.12",
-    "eslint": "^8.57.1",
+    "@vue/eslint-config-typescript": "^14.1.3",
+    "eslint": "^9.14.0",
+    "eslint-plugin-vue": "^9.31.0",
     "lint-staged": "^15.2.10",
     "prettier": "^3.3.3",
-    "simple-git-hooks": "^2.11.1"
+    "simple-git-hooks": "^2.11.1",
+    "svelte": "^5.1.15"
   },
   "config": {
     "commitizen": {
