@@ -1,10 +1,10 @@
 /** @type {import("prettier").Config} */
-module.exports = {
+export default {
   printWidth: 179,
   tabWidth: 2,
   useTabs: false,
   quoteProps: 'as-needed',
-  vueIndentScriptAndStyle: true,
+  vueIndentScriptAndStyle: false,
   semi: false,
   singleQuote: true,
   trailingComma: 'none',
