@@ -1,15 +1,17 @@
-const { ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS } = require('./constants')
+import ImportPlugin from 'eslint-plugin-import'
+import { EXTENSIONS } from './constants.js'
 
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  overrides: [
-    {
-      files: ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS,
-      plugins: [ 'import' ],
-      rules: {
-        'import/no-extraneous-dependencies': 'error',
-        'import/exports-last': 'error'
-      }
+/** @type {import("eslint").Linter.Config[]} */
+export default [
+  {
+    name: 'cenk1cenk2/import/strict',
+    files: [...EXTENSIONS],
+    plugins: {
+      import: ImportPlugin
+    },
+    rules: {
+      'import/no-extraneous-dependencies': 'error',
+      'import/exports-last': 'error'
     }
-  ]
-}
+  }
+]
