@@ -1,10 +1,12 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: [ './index' ],
-  overrides: [
-    {
-      files: [ '*.jsx', '*.tsx' ],
-      plugins: [ 'react', 'react-hooks' ],
+import js from './js.js'
+
+/** @type {import("eslint").Linter.Config[]} */
+export default [
+  ...js,
+  {
+    files: ['**/*.jsx', '**/*.tsx'],
+    plugins: ['react', 'react-hooks'],
+    languageOptions: {
       parserOptions: {
         project: null,
         ecmaFeatures: {
@@ -12,55 +14,55 @@ module.exports = {
         },
         ecmaVersion: 2018,
         sourceType: 'module'
-      },
-      settings: {
-        react: {
-          version: 'detect'
-        }
-      },
-      rules: {
-        'no-unused-vars': 'off',
-        'no-undef': 'off',
-        indent: 'off',
-        'react/jsx-indent': [ 'error', 2 ],
-        'react/jsx-indent-props': [ 'error', 2 ],
-        'react/jsx-curly-newline': [ 'error', 'consistent' ],
-        'react/jsx-fragments': [ 'error', 'element' ],
-        'react/require-render-return': 'error',
-        'react/self-closing-comp': [
-          'error',
-          {
-            component: true,
-            html: true
-          }
-        ],
-        'react/jsx-tag-spacing': [
-          'error',
-          {
-            closingSlash: 'never',
-            beforeSelfClosing: 'always',
-            afterOpening: 'never',
-            beforeClosing: 'never'
-          }
-        ],
-        'react/jsx-closing-tag-location': 'error',
-        'react/jsx-curly-brace-presence': [ 'error', { props: 'never', children: 'never' } ],
-        'react/jsx-equals-spacing': [ 'error', 'never' ],
-        'react/jsx-first-prop-new-line': [ 'error', 'multiline' ],
-        'react/jsx-props-no-multi-spaces': 'error',
-        'react/jsx-wrap-multilines': [
-          'error',
-          {
-            declaration: 'parens-new-line',
-            assignment: 'parens-new-line',
-            return: 'parens-new-line',
-            arrow: 'parens-new-line',
-            condition: 'ignore',
-            logical: 'ignore',
-            prop: 'ignore'
-          }
-        ]
       }
+    },
+    settings: {
+      react: {
+        version: 'detect'
+      }
+    },
+    rules: {
+      'no-unused-vars': 'off',
+      'no-undef': 'off',
+      indent: 'off',
+      'react/jsx-indent': ['error', 2],
+      'react/jsx-indent-props': ['error', 2],
+      'react/jsx-curly-newline': ['error', 'consistent'],
+      'react/jsx-fragments': ['error', 'element'],
+      'react/require-render-return': 'error',
+      'react/self-closing-comp': [
+        'error',
+        {
+          component: true,
+          html: true
+        }
+      ],
+      'react/jsx-tag-spacing': [
+        'error',
+        {
+          closingSlash: 'never',
+          beforeSelfClosing: 'always',
+          afterOpening: 'never',
+          beforeClosing: 'never'
+        }
+      ],
+      'react/jsx-closing-tag-location': 'error',
+      'react/jsx-curly-brace-presence': ['error', { props: 'never', children: 'never' } ],
+      'react/jsx-equals-spacing': ['error', 'never'],
+      'react/jsx-first-prop-new-line': ['error', 'multiline'],
+      'react/jsx-props-no-multi-spaces': 'error',
+      'react/jsx-wrap-multilines': [
+        'error',
+        {
+          declaration: 'parens-new-line',
+          assignment: 'parens-new-line',
+          return: 'parens-new-line',
+          arrow: 'parens-new-line',
+          condition: 'ignore',
+          logical: 'ignore',
+          prop: 'ignore'
+        }
+      ]
     }
-  ]
-}
+  }
+]
