@@ -1,15 +1,14 @@
-const { loadTsConfig } = require('load-tsconfig')
-
-const { ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS } = require('./constants')
+import { loadTsConfig } from 'load-tsconfig'
+import ImportPlugin from 'eslint-plugin-import'
 
 /**
- * @param {{ paths?: string[] tsconfig?: string, tsconfigDir?: string}} options
+ * @param {{ paths?: string[], tsconfig?: string, tsconfigDir?: string}} options
  * @returns {import("eslint").Linter.Config['rules']}
  */
-function generateImportGroups (options) {
-  const defaults = [ '@cenk1cenk2', '@cenk1cenk2-*' ]
+export function generateImportGroups(options) {
+  const defaults = ['@cenk1cenk2', '@cenk1cenk2-*']
 
-  options.paths = Array.isArray(options.paths) ? [ ...options.paths, ...defaults ] : defaults
+  options.paths = Array.isArray(options.paths) ? [...options.paths, ...defaults] : defaults
 
   /** @type {string[]} */
   let tsConfigPaths = []
@@ -55,10 +54,10 @@ function generateImportGroups (options) {
             ]
           }, [])
         ],
-        pathGroupsExcludedImportTypes: [ 'builtin' ],
+        pathGroupsExcludedImportTypes: ['builtin'],
         groups: [
-          [ 'builtin', 'external' ],
-          [ 'index', 'parent', 'sibling' ]
+          ['builtin', 'external'],
+          ['index', 'parent', 'sibling']
         ],
         'newlines-between': 'always',
         alphabetize: {
@@ -71,17 +70,16 @@ function generateImportGroups (options) {
 }
 
 /**
- * @param {{ paths?: string[] tsconfig?: string, tsconfigDir?: string}} options
- * @returns {import("eslint").Linter.ConfigOverride}
+ * @param {{ paths?: string[], tsconfig?: string, tsconfigDir?: string}} options
+ * @returns {import("eslint").Linter.Config[]}
  */
-function generateImportGroupsWithOverride (options) {
+export function configImportGroup(options) {
   return [
     {
-      files: ALL_KNOWN_JAVASCRIPT_FILE_EXTENSIONS,
-      plugins: [ 'import' ],
+      plugins: {
+        import: ImportPlugin
+      },
       rules: generateImportGroups(options)
     }
   ]
 }
-
-module.exports = { generateImportGroups, generateImportGroupsWithOverride }
