@@ -1,21 +1,28 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: [ 'plugin:svelte/recommended', './typescript-dynamic' ],
-  parserOptions: {
-    extraFileExtensions: [ '.svelte' ]
-  },
-  overrides: [
-    {
-      files: [ '*.svelte' ],
-      parser: 'svelte-eslint-parser',
+import TypescriptParser from '@typescript-eslint/parser'
+import SveltePlugin from 'eslint-plugin-svelte'
+import SvelteParser from 'svelte-eslint-parser'
+
+import { EXTENSIONS_SVELTE } from './constants.js'
+import ConfigTypescript from './typescript-dynamic.js'
+
+/** @type {import("eslint").Linter.Config[]} */
+export default [
+  ...SveltePlugin.configs['flat/recommended'],
+
+  ...ConfigTypescript,
+
+  {
+    name: 'cenk1cenk2/svelte/typescript/recommended',
+    files: [...EXTENSIONS_SVELTE],
+    languageOptions: {
+      parser: SvelteParser,
       parserOptions: {
-        parser: '@typescript-eslint/parser'
-      },
-      rules: {
-        'max-len': 'off',
-        'no-throw-literal': 'off',
-        '@typescript-eslint/no-throw-literal': 'off'
+        parser: TypescriptParser,
+        extraFileExtensions: ['.svelte']
       }
+    },
+    rules: {
+      'max-len': 'off'
     }
-  ]
-}
+  }
+]
